+++ test/quick/quicktest.cpp
@@ -25,7 +25,7 @@ int main() {
     if (result.size() != data.size())
         success = false;
 
-    for (int index = 0; index < ((int)result.size() - 1); index++) {
+    for (long int index = 0; index < ((long int)result.size() - 1); index++) {
         if (result[index+1] < result[index]) success = false;
     }
